requirements.txt
web3==6.15.1
py-solc-x==2.0.0
py-ecc==8.0.0
eth-account==0.10.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
seaborn==0.13.2
tqdm==4.66.4
cryptography==42.0.8
scipy==1.13.1
